backend/models/bhargavi_models.py
"""
Bhargavi Model Integration
Models from: Bhargavidev26/Early-Tomato-Leaf-Disease-Prediction

Contains:
- Custom CNN model (94%/93% accuracy)
- VGG16 Transfer Learning (24%/24% accuracy)  
- ResNet50 Transfer Learning (68%/73% accuracy)
"""

import tensorflow as tf
from tensorflow import keras
from tensorflow.keras import layers
import numpy as np
from PIL import Image
import io


class BhargaviTomatoDiseaseClassifier:
    """Tomato Disease Classification using models from Bhargavi repository"""
    
    def __init__(self, model_type='cnn'):
        """
        Initialize the classifier
        
        Args:
            model_type (str): Type of model to use ('cnn', 'vgg16', 'resnet50')
        """
        self.model_type = model_type
        self.model = None
        self.class_names = [
            'Tomato_Bacterial_spot',
            'Tomato_Early_blight', 
            'Tomato_Late_blight',
            'Tomato_Leaf_Mold',
            'Tomato_Septoria_leaf_spot',
            'Tomato_Spider_mites_Two_spotted_spider_mite',
            'Tomato__Target_Spot',
            'Tomato__Tomato_YellowLeaf__Curl_Virus',
            'Tomato__Tomato_mosaic_virus',
            'Tomato_healthy'
        ]
        self.img_height = 224
        self.img_width = 224
        
    def build_cnn_model(self):
        """Build custom CNN model based on Bhargavi's implementation"""
        model = keras.Sequential([
            layers.Conv2D(32, (3, 3), activation='relu', input_shape=(self.img_height, self.img_width, 3)),
            layers.MaxPooling2D((2, 2)),
            layers.Conv2D(64, (3, 3), activation='relu'),
            layers.MaxPooling2D((2, 2)),
            layers.Conv2D(64, (3, 3), activation='relu'),
            layers.MaxPooling2D((2, 2)),
            layers.Conv2D(128, (3, 3), activation='relu'),
            layers.MaxPooling2D((2, 2)),
            layers.Conv2D(128, (3, 3), activation='relu'),
            layers.MaxPooling2D((2, 2)),
            layers.Flatten(),
            layers.Dropout(0.5),
            layers.Dense(512, activation='relu'),
            layers.Dense(len(self.class_names), activation='softmax')
        ])
        
        model.compile(
            optimizer='adam',
            loss='categorical_crossentropy',
            metrics=['accuracy']
        )
        
        return model
    
    def build_vgg16_model(self):
        """Build VGG16 transfer learning model"""
        base_model = tf.keras.applications.VGG16(
            weights='imagenet',
            include_top=False,
            input_shape=(self.img_height, self.img_width, 3)
        )
        
        base_model.trainable = False
        
        model = keras.Sequential([
            base_model,
            layers.GlobalAveragePooling2D(),
            layers.Dropout(0.2),
            layers.Dense(len(self.class_names), activation='softmax')
        ])
        
        model.compile(
            optimizer='adam',
            loss='categorical_crossentropy',
            metrics=['accuracy']
        )
        
        return model
    
    def build_resnet50_model(self):
        """Build ResNet50 transfer learning model"""
        base_model = tf.keras.applications.ResNet50(
            weights='imagenet',
            include_top=False,
            input_shape=(self.img_height, self.img_width, 3)
        )
        
        base_model.trainable = False
        
        model = keras.Sequential([
            base_model,
            layers.GlobalAveragePooling2D(),
            layers.Dropout(0.2),
            layers.Dense(len(self.class_names), activation='softmax')
        ])
        
        model.compile(
            optimizer='adam',
            loss='categorical_crossentropy',
            metrics=['accuracy']
        )
        
        return model
    
    def load_model(self):
        """Load the specified model"""
        if self.model_type == 'cnn':
            self.model = self.build_cnn_model()
        elif self.model_type == 'vgg16':
            self.model = self.build_vgg16_model()
        elif self.model_type == 'resnet50':
            self.model = self.build_resnet50_model()
        else:
            raise ValueError(f"Unknown model type: {self.model_type}")
        
        return self.model
    
    def preprocess_image(self, image_data):
        """
        Preprocess image for prediction
        
        Args:
            image_data: Image data (PIL Image or bytes)
            
        Returns:
            Preprocessed image array
        """
        if isinstance(image_data, bytes):
            image = Image.open(io.BytesIO(image_data))
        else:
            image = image_data
            
        # Convert to RGB if necessary
        if image.mode != 'RGB':
            image = image.convert('RGB')
            
        # Resize image
        image = image.resize((self.img_width, self.img_height))
        
        # Convert to numpy array and normalize
        img_array = tf.keras.utils.img_to_array(image)
        img_array = tf.expand_dims(img_array, 0)  # Create batch dimension
        img_array = img_array / 255.0  # Normalize to [0,1]
        
        return img_array
    
    def predict(self, image_data):
        """
        Make prediction on input image
        
        Args:
            image_data: Image data (PIL Image or bytes)
            
        Returns:
            Dictionary with prediction results
        """
        if self.model is None:
            self.load_model()
            
        # Preprocess image
        processed_image = self.preprocess_image(image_data)
        
        # Make prediction
        predictions = self.model.predict(processed_image)
        
        # Get top prediction
        predicted_class_idx = np.argmax(predictions[0])
        confidence = float(predictions[0][predicted_class_idx])
        predicted_class = self.class_names[predicted_class_idx]
        
        # Get all predictions
        all_predictions = []
        for i, prob in enumerate(predictions[0]):
            all_predictions.append({
                'class': self.class_names[i],
                'confidence': float(prob)
            })
        
        # Sort by confidence
        all_predictions.sort(key=lambda x: x['confidence'], reverse=True)
        
        return {
            'predicted_class': predicted_class,
            'confidence': confidence,
            'all_predictions': all_predictions[:5],  # Top 5 predictions
            'model_type': self.model_type
        }


def get_model_info():
    """Get information about available models"""
    return {
        'cnn': {
            'name': 'Custom CNN',
            'accuracy': '94%/93%',
            'description': 'Custom CNN architecture optimized for tomato disease classification'
        },
        'vgg16': {
            'name': 'VGG16 Transfer Learning',
            'accuracy': '24%/24%',
            'description': 'VGG16 pre-trained model with custom classifier for tomato diseases'
        },
        'resnet50': {
            'name': 'ResNet50 Transfer Learning', 
            'accuracy': '68%/73%',
            'description': 'ResNet50 pre-trained model with custom classifier for tomato diseases'
        }
    }