backend/requirements.txt
flask
flask-cors
pillow
numpy
tensorflow
fastapi
uvicorn[standard]
python-multipart
pydantic